+++ crates/menmos-interface/src/directory/dir.rs
@@ -126,7 +126,7 @@ impl Query {
     }
 
     #[must_use]
-    pub fn and_meta<K: Into<String>, V: Into<String>>(mut self, k: K, v: V) -> Self {
+    pub fn and_field<K: Into<String>, V: Into<String>>(mut self, k: K, v: V) -> Self {
         let new_expr = Expression::Field(ExpressionField::Field {
             key: k.into(),
             value: v.into(),
